+++ crates/ryft-xla-sys/src/c++/mlir/dialects/nvgpu.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "../../common.h"
+
 #include <stdbool.h>
 #include <stdint.h>
 
@@ -17,34 +19,40 @@ enum MlirNvgpuEnumAttribute {
   RYFT_MLIR_NVGPU_ENUM_ATTRIBUTE_RCP_ROUNDING_MODE = 4,
 };
 
-bool mlirTypeIsANvgpuDeviceAsyncTokenType(MlirType type);
-MlirType mlirNvgpuDeviceAsyncTokenTypeGet(MlirContext context);
+RYFT_XLA_SYS_EXPORT bool mlirTypeIsANvgpuDeviceAsyncTokenType(MlirType type);
+RYFT_XLA_SYS_EXPORT MlirType mlirNvgpuDeviceAsyncTokenTypeGet(MlirContext context);
 
-bool mlirTypeIsANvgpuMBarrierGroupType(MlirType type);
-MlirType mlirNvgpuMBarrierGroupTypeGet(MlirContext context, MlirAttribute memory_space, uint32_t num_barriers);
-MlirAttribute mlirNvgpuMBarrierGroupTypeGetMemorySpace(MlirType type);
-uint32_t mlirNvgpuMBarrierGroupTypeGetNumBarriers(MlirType type);
+RYFT_XLA_SYS_EXPORT bool mlirTypeIsANvgpuMBarrierGroupType(MlirType type);
+RYFT_XLA_SYS_EXPORT MlirType mlirNvgpuMBarrierGroupTypeGet(
+    MlirContext context,
+    MlirAttribute memory_space,
+    uint32_t num_barriers);
+RYFT_XLA_SYS_EXPORT MlirAttribute mlirNvgpuMBarrierGroupTypeGetMemorySpace(MlirType type);
+RYFT_XLA_SYS_EXPORT uint32_t mlirNvgpuMBarrierGroupTypeGetNumBarriers(MlirType type);
 
-bool mlirTypeIsANvgpuMBarrierTokenType(MlirType type);
-MlirType mlirNvgpuMBarrierTokenTypeGet(MlirContext context);
+RYFT_XLA_SYS_EXPORT bool mlirTypeIsANvgpuMBarrierTokenType(MlirType type);
+RYFT_XLA_SYS_EXPORT MlirType mlirNvgpuMBarrierTokenTypeGet(MlirContext context);
 
-MlirType mlirNvgpuTensorMapDescriptorTypeGetTensor(MlirType type);
-uint32_t mlirNvgpuTensorMapDescriptorTypeGetSwizzle(MlirType type);
-uint32_t mlirNvgpuTensorMapDescriptorTypeGetL2Promo(MlirType type);
-uint32_t mlirNvgpuTensorMapDescriptorTypeGetOob(MlirType type);
-uint32_t mlirNvgpuTensorMapDescriptorTypeGetInterleave(MlirType type);
+RYFT_XLA_SYS_EXPORT MlirType mlirNvgpuTensorMapDescriptorTypeGetTensor(MlirType type);
+RYFT_XLA_SYS_EXPORT uint32_t mlirNvgpuTensorMapDescriptorTypeGetSwizzle(MlirType type);
+RYFT_XLA_SYS_EXPORT uint32_t mlirNvgpuTensorMapDescriptorTypeGetL2Promo(MlirType type);
+RYFT_XLA_SYS_EXPORT uint32_t mlirNvgpuTensorMapDescriptorTypeGetOob(MlirType type);
+RYFT_XLA_SYS_EXPORT uint32_t mlirNvgpuTensorMapDescriptorTypeGetInterleave(MlirType type);
 
-bool mlirTypeIsANvgpuWarpgroupMatrixDescriptorType(MlirType type);
-MlirType mlirNvgpuWarpgroupMatrixDescriptorTypeGet(MlirContext context, MlirType tensor);
-MlirType mlirNvgpuWarpgroupMatrixDescriptorTypeGetTensor(MlirType type);
+RYFT_XLA_SYS_EXPORT bool mlirTypeIsANvgpuWarpgroupMatrixDescriptorType(MlirType type);
+RYFT_XLA_SYS_EXPORT MlirType mlirNvgpuWarpgroupMatrixDescriptorTypeGet(MlirContext context, MlirType tensor);
+RYFT_XLA_SYS_EXPORT MlirType mlirNvgpuWarpgroupMatrixDescriptorTypeGetTensor(MlirType type);
 
-bool mlirTypeIsANvgpuWarpgroupAccumulatorType(MlirType type);
-MlirType mlirNvgpuWarpgroupAccumulatorTypeGet(MlirContext context, MlirType fragmented);
-MlirType mlirNvgpuWarpgroupAccumulatorTypeGetFragmented(MlirType type);
+RYFT_XLA_SYS_EXPORT bool mlirTypeIsANvgpuWarpgroupAccumulatorType(MlirType type);
+RYFT_XLA_SYS_EXPORT MlirType mlirNvgpuWarpgroupAccumulatorTypeGet(MlirContext context, MlirType fragmented);
+RYFT_XLA_SYS_EXPORT MlirType mlirNvgpuWarpgroupAccumulatorTypeGetFragmented(MlirType type);
 
-bool mlirAttributeIsANvgpuEnumAttr(MlirAttribute attribute, enum MlirNvgpuEnumAttribute kind);
-MlirAttribute mlirNvgpuEnumAttrGet(MlirContext context, enum MlirNvgpuEnumAttribute kind, uint32_t value);
-uint32_t mlirNvgpuEnumAttrGetValue(MlirAttribute attribute, enum MlirNvgpuEnumAttribute kind);
+RYFT_XLA_SYS_EXPORT bool mlirAttributeIsANvgpuEnumAttr(MlirAttribute attribute, enum MlirNvgpuEnumAttribute kind);
+RYFT_XLA_SYS_EXPORT MlirAttribute mlirNvgpuEnumAttrGet(
+    MlirContext context,
+    enum MlirNvgpuEnumAttribute kind,
+    uint32_t value);
+RYFT_XLA_SYS_EXPORT uint32_t mlirNvgpuEnumAttrGetValue(MlirAttribute attribute, enum MlirNvgpuEnumAttribute kind);
 
 #ifdef __cplusplus
 }